+++ module/item/item-attack.mjs
@@ -2198,6 +2198,7 @@ export async function _onApplyDamageToSpecificToken(event, tokenId) {
             token,
             damageDetail,
             defense,
+            defenseTags,
         );
     }
     const senseAffecting = getPowerInfo({
@@ -2494,6 +2495,7 @@ async function _onApplyAdjustmentToSpecificToken(
     token,
     damageDetail,
     defense,
+    defenseTags,
 ) {
     if (
         adjustmentItem.actor.id === token.actor.id &&
@@ -2523,6 +2525,8 @@ async function _onApplyAdjustmentToSpecificToken(
         );
     }
 
+    const adjustmentItemTags = getAttackTags(adjustmentItem);
+
     const reductionChatMessages = [];
     const reductionTargetActor = token.actor;
     for (const reduce of reducesArray) {
@@ -2539,7 +2543,11 @@ async function _onApplyAdjustmentToSpecificToken(
         );
     }
     if (reductionChatMessages.length > 0) {
-        await renderAdjustmentChatCards(reductionChatMessages);
+        await renderAdjustmentChatCards(
+            reductionChatMessages,
+            adjustmentItemTags,
+            defenseTags,
+        );
     }
 
     const enhancementChatMessages = [];
@@ -2563,7 +2571,11 @@ async function _onApplyAdjustmentToSpecificToken(
         );
     }
     if (enhancementChatMessages.length > 0) {
-        await renderAdjustmentChatCards(enhancementChatMessages);
+        await renderAdjustmentChatCards(
+            enhancementChatMessages,
+            adjustmentItemTags,
+            [], // don't show any defense tags as this is an enhancement adjustment
+        );
     }
 }